review comments: docstring & simpler map_vals
